requirements.txt
Django==1.11.7
djangorestframework==3.7.3
django_extensions==1.9.7
coinbase==2.0.6
psycopg2==2.7.3.2
celery==4.1.0
django-cors-headers==2.1.0
